Add SteadyFormulation for generalised time independent solves
