+++ src/formulations/transient_formulation.cpp
@@ -15,6 +15,43 @@ GetTimeDerivativeNames(std::vector<std::string> gridfunction_names) {
   return time_derivative_names;
 }
 
+void TimeDomainEquationSystemOperator::SetVariables() {
+  state_var_names = _equation_system->var_names;
+  local_test_vars = populateVectorFromNamedFieldsMap<mfem::ParGridFunction>(
+      _variables, _equation_system->var_names);
+  local_trial_vars = populateVectorFromNamedFieldsMap<mfem::ParGridFunction>(
+      _variables, _equation_system->var_time_derivative_names);
+
+  // Set operator size and block structure
+  block_trueOffsets.SetSize(local_test_vars.size() + 1);
+  block_trueOffsets[0] = 0;
+  for (unsigned int ind = 0; ind < local_test_vars.size(); ++ind) {
+    block_trueOffsets[ind + 1] =
+        local_test_vars.at(ind)->ParFESpace()->TrueVSize();
+  }
+  block_trueOffsets.PartialSum();
+
+  true_offsets.SetSize(local_test_vars.size() + 1);
+  true_offsets[0] = 0;
+  for (unsigned int ind = 0; ind < local_test_vars.size(); ++ind) {
+    true_offsets[ind + 1] = local_test_vars.at(ind)->ParFESpace()->GetVSize();
+  }
+  true_offsets.PartialSum();
+
+  this->height = true_offsets[local_test_vars.size()];
+  this->width = true_offsets[local_test_vars.size()];
+  trueX.Update(block_trueOffsets);
+  trueRhs.Update(block_trueOffsets);
+
+  // Populate vector of active auxiliary variables
+  active_aux_var_names.resize(0);
+  for (auto &aux_var_name : aux_var_names) {
+    if (_variables.Has(aux_var_name)) {
+      active_aux_var_names.push_back(aux_var_name);
+    }
+  }
+};
+
 void TimeDomainEquationSystemOperator::Init(mfem::Vector &X) {
   // Define material property coefficients
   for (unsigned int ind = 0; ind < local_test_vars.size(); ++ind) {
@@ -57,43 +94,6 @@ void TimeDomainEquationSystemOperator::SetEquationSystem(
     hephaestus::TimeDependentEquationSystem *equation_system) {
   _equation_system = equation_system;
 }
-
-void TimeDomainEquationSystemOperator::SetVariables() {
-  state_var_names = _equation_system->var_names;
-  local_test_vars = populateVectorFromNamedFieldsMap<mfem::ParGridFunction>(
-      _variables, _equation_system->var_names);
-  local_trial_vars = populateVectorFromNamedFieldsMap<mfem::ParGridFunction>(
-      _variables, _equation_system->var_time_derivative_names);
-
-  // Set operator size and block structure
-  block_trueOffsets.SetSize(local_test_vars.size() + 1);
-  block_trueOffsets[0] = 0;
-  for (unsigned int ind = 0; ind < local_test_vars.size(); ++ind) {
-    block_trueOffsets[ind + 1] =
-        local_test_vars.at(ind)->ParFESpace()->TrueVSize();
-  }
-  block_trueOffsets.PartialSum();
-
-  true_offsets.SetSize(local_test_vars.size() + 1);
-  true_offsets[0] = 0;
-  for (unsigned int ind = 0; ind < local_test_vars.size(); ++ind) {
-    true_offsets[ind + 1] = local_test_vars.at(ind)->ParFESpace()->GetVSize();
-  }
-  true_offsets.PartialSum();
-
-  this->height = true_offsets[local_test_vars.size()];
-  this->width = true_offsets[local_test_vars.size()];
-  trueX.Update(block_trueOffsets);
-  trueRhs.Update(block_trueOffsets);
-
-  // Populate vector of active auxiliary variables
-  active_aux_var_names.resize(0);
-  for (auto &aux_var_name : aux_var_names) {
-    if (_variables.Has(aux_var_name)) {
-      active_aux_var_names.push_back(aux_var_name);
-    }
-  }
-};
 
 TransientFormulation::TransientFormulation()
     : equation_system(NULL), td_operator(NULL), oneCoef(1.0){};
